Updated patch to test cleanly against master
